Move GenericBFS to its own package
